Add files via upload
The cards in my project are represented with two a + b * 0.1. a and b are integers. a is the number of the card, which ranges from 1 to 13. b is the suit of the card, which ranges from 1 to 4 (1 = hearts, 2 = spades, 3 = diamonds, 4 = clubs).

My initial generation contains 100 hands which I randomly generated with a stack of 52 cards. The loop runs for a total of 250 generations. The fitness of each hand in each generation is evaluated with the fitness function. For each generation, the hands who have a higher rank than full house are thrown out before selection for generating next generatinon. Before elimination by ranks, it rarely happens that a high rank hand appears among the 100 or so hands in each generation. The reason for that is the probability of having i.e. four of a kind is very very small compared to regular hands. Through cross-over, mutation, the average fitness score does increase. However, it's not a drastic increase as I anticipated. The average fitness score of the